Update address resolver
